project restruturation
