--- MusicDownloader/MusicDownloader.http
@@ -1,6 +1,0 @@
-@MusicDownloader_HostAddress = http://localhost:5217
-
-GET {{MusicDownloader_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
