--- MusicDownloader/Program.cs
@@ -1,29 +1,0 @@
-using MusicDownloader.Services.Downloader;
-using MusicDownloader.Services.IOManagement;
-
-var builder = WebApplication.CreateBuilder(args);
-
-// Add services to the container.
-
-builder.Services.AddControllers();
-// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-builder.Services.AddScoped<Downloader>(x => new (builder.Configuration.GetValue<string>("BaseUrl")));
-builder.Services.AddScoped<FileManager>(x => new (builder.Configuration.GetValue<string>("DefaultFolderName")));
-
-var app = builder.Build();
-
-// Configure the HTTP request pipeline.
-if (app.Environment.IsDevelopment()) {
-    app.UseSwagger();
-    app.UseSwaggerUI();
-}
-
-app.UseHttpsRedirection();
-
-app.UseAuthorization();
-
-app.MapControllers();
-
-app.Run();
