--- MusicDownloader/Services/Downloader/DownloadException.cs
@@ -1,7 +1,0 @@
-﻿namespace MusicDownloader.Services.Downloader {
-    public class DownloadException : Exception{
-        public  DownloadException(string message) : base(message) {
-
-        }
-    }
-}
