--- MusicDownloader/Services/Downloader/GetVideoInfoRequest.cs
@@ -1,6 +1,0 @@
-﻿namespace MusicDownloader.Services.Downloader {
-    public class GetVideoInfoRequest {
-        public string url {  get; set; }
-        public int videoSource { get; set; }
-    }
-}
