--- MusicDownloader/Services/Downloader/Video.cs
@@ -1,6 +1,0 @@
-﻿namespace MusicDownloader.Services.Downloader {
-    public class Video {
-        public string Title { get; set; }
-        public byte[] Data { get; set; }
-    }
-}
