--- WebMVC/Pages/Index.cshtml
@@ -1,6 +1,0 @@
-﻿@page
-@model IndexModel
-
-<div class="text-center">
-    URL: <input type="text"/>
-</div>
