--- WebMVC/Pages/Index.cshtml.cs
@@ -1,16 +1,0 @@
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace WebMVC.Pages {
-    public class IndexModel : PageModel {
-        private readonly ILogger<IndexModel> _logger;
-
-        public IndexModel(ILogger<IndexModel> logger) {
-            _logger = logger;
-        }
-
-        public void OnGet() {
-
-        }
-    }
-}
