--- WebMVC/Pages/Settings.cshtml
@@ -1,4 +1,0 @@
-﻿@page
-<h1>Settings</h1>
-
-<p>This page will be the application configuration page, and is going to be implemented in the future (maybe).</p>
