--- WebMVC/Pages/Settings.cshtml.cs
@@ -1,16 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace WebMVC.Pages {
-    public class PrivacyModel : PageModel {
-        private readonly ILogger<PrivacyModel> _logger;
-
-        public PrivacyModel(ILogger<PrivacyModel> logger) {
-            _logger = logger;
-        }
-
-        public void OnGet() {
-        }
-    }
-
-}
