+++ WebTools/Models/ErrorViewModel.cs
@@ -1,0 +1,7 @@
+namespace WebTools.Models {
+    public class ErrorViewModel {
+        public string? RequestId { get; set; }
+
+        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+    }
+}
