+++ WebTools/Views/Home/Index.cshtml
@@ -1,0 +1,8 @@
+﻿
+<div class="text-center">
+    <form asp-action="DownloadVideo" method="get" id="formDownload">
+        URL: <input type="text" class="" id="videoUrl" name="url" />
+        <input type="submit" downloadVideoButton" value="Download" />
+        <div id="downloadResult"></div>
+    </form>
+</div>
