+++ WebTools/Views/Shared/Error.cshtml
@@ -1,5 +1,4 @@
-﻿@page
-@model ErrorModel
+﻿@model ErrorViewModel
 @{
     ViewData["Title"] = "Error";
 }
@@ -16,7 +15,7 @@
 
 <h3>Development Mode</h3>
 <p>
-    Swapping to the <strong>Development</strong> environment displays detailed information about the error that occurred.
+    Swapping to <strong>Development</strong> environment will display more detailed information about the error that occurred.
 </p>
 <p>
     <strong>The Development environment shouldn't be enabled for deployed applications.</strong>
