+++ WebTools/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
-﻿@using WebMVC
-@namespace WebMVC.Pages
+﻿@using WebTools
+@using WebTools.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
