+++ WebTools/wwwroot/lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.js
