+++ WebTools/wwwroot/lib/jquery-validation/LICENSE.md
