+++ WebTools/wwwroot/lib/jquery/LICENSE.txt
